Moves enumeration utilities into the package
