+++ src/algmatch/utils/enumerators/HR/hrGenericBruteForcer.py
