+++ src/algmatch/utils/enumerators/SPAS/spasGenericBruteForcer.py
