+++ src/algmatch/utils/enumerators/SR/srEnumerator.py
