+++ tests/SMTests/SM/smVerifier.py
@@ -2,7 +2,7 @@ from algmatch.stableMarriageProblem import StableMarriageProblem
 
 from tests.abstractTestClasses.abstractVerifier import AbstractVerifier
 from algmatch.utils.generators.SM.smGenerator import SMGenerator
-from tests.SMTests.utils.smEnumerator import SMEnumerator
+from algmatch.utils.enumerators.SM.smEnumerator import SMEnumerator
 
 
 class SMVerifier(AbstractVerifier):
