+++ tests/SMTests/SMTStrong/smtStrongVerifier.py
@@ -2,7 +2,7 @@ from algmatch.stableMarriageProblemWithTies import StableMarriageProblemWithTies
 
 from tests.abstractTestClasses.abstractVerifier import AbstractVerifier
 from algmatch.utils.generators.SM.smtGenerator import SMTGenerator
-from tests.SMTests.utils.smtEnumerator import SMTEnumerator
+from algmatch.utils.enumerators.SM.smtEnumerator import SMTEnumerator
 
 
 class SMTStrongVerifier(AbstractVerifier):
